# postman/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API Raizes do Nordeste",
        "description": "Back-end da rede de lanchonetes Raizes do Nordeste.\n\nAtende varias unidades da franquia e varios canais de venda\n(APP, TOTEM, BALCAO, PICKUP e WEB).\n\nComo testar:\n1. Chame POST /auth/login com um usuario do seed.\n2. Clique em Authorize aqui em cima e cole o accessToken.\n3. As rotas protegidas passam a funcionar.\n\nUsuarios do seed (senha Senha@123):\n- admin@raizes.com.br (ADMIN)\n- gerente.recife@raizes.com.br (GERENTE)\n- atendente.recife@raizes.com.br (ATENDENTE)\n- cozinha.recife@raizes.com.br (COZINHA)\n- cliente@exemplo.com (CLIENTE)",
        "contact": {
            "name": "Gean Brandao"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Ambiente local"
        },
        {
            "url": "/",
            "description": "Mesma origem que serviu esta pagina"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
    "paths": {
        "/unidades/{unidadeId}": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Detalhar unidade",
                "description": "Dados de uma unidade pelo id. Rota publica: o cliente precisa saber onde a loja fica antes de pedir.",
                "operationId": "detalhar",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade n\u00e3o encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            },
            "put": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Atualizar unidade",
                "description": "Exclusivo do perfil ADMIN.",
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnidadeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Unidade atualizada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade n\u00e3o encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/produtos/{produtoId}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Detalhar produto",
                "description": "Dados de um produto do catalogo da rede. Traz tambem o que foi inativado, com ativo = false, porque pedido antigo aponta para ele.",
                "operationId": "detalhar_1",
                "parameters": [
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdutoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto",
                "description": "Perfis ADMIN e GERENTE.",
                "operationId": "atualizar_1",
                "parameters": [
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdutoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Inativar produto",
                "description": "Exclusivo do perfil ADMIN. Marca como inativo em vez de apagar, para n\u00e3o quebrar os pedidos antigos que apontam para o produto.",
                "operationId": "inativar",
                "parameters": [
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto inativado"
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos",
                "description": "Catalogo da rede, com filtro opcional por categoria.",
                "operationId": "listar_1",
                "parameters": [
                    {
                        "name": "categoria",
                        "in": "query",
                        "description": "Filtra por categoria, ex.: TAPIOCA",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginaResponseProdutoResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "N\u00e3o autenticado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Cadastrar produto",
                "description": "Perfis ADMIN e GERENTE.",
                "operationId": "criar_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Produto cadastrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdutoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos invalidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/pedidos/{pedidoId}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Detalhar pedido",
                "description": "Pedido que n\u00e3o pertence a quem esta consultando devolve 404, n\u00e3o 403: dizer 'existe mas n\u00e3o e seu' ja confirmaria a existencia dele.",
                "operationId": "detalhar_2",
                "parameters": [
                    {
                        "name": "pedidoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido n\u00e3o encontrado ou n\u00e3o visivel a esse perfil",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/pagamentos/{pagamentoId}": {
            "get": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Consultar pagamento",
                "description": "A visibilidade e a mesma do pedido: quem n\u00e3o ve o pedido n\u00e3o ve quanto foi cobrado nele.",
                "operationId": "detalhar_3",
                "parameters": [
                    {
                        "name": "pagamentoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pagamento n\u00e3o encontrado ou n\u00e3o visivel",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UnidadeRequest": {
                "type": "object",
                "description": "Dados de uma unidade",
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Raizes Boa Viagem",
                        "maxLength": 120,
                        "minLength": 3
                    },
                    "cidade": {
                        "type": "string",
                        "example": "Recife",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "uf": {
                        "type": "string",
                        "example": "PE",
                        "minLength": 1,
                        "pattern": "^[A-Za-z]{2}$"
                    },
                    "tipoOperacao": {
                        "type": "string",
                        "description": "COMPLETA prepara o cardapio inteiro; REDUZIDA prepara so parte dele",
                        "enum": [
                            "COMPLETA",
                            "REDUZIDA"
                        ],
                        "example": "COMPLETA"
                    }
                },
                "required": [
                    "cidade",
                    "nome",
                    "tipoOperacao",
                    "uf"
                ]
            },
            "ErrorDetail": {
                "type": "object",
                "description": "Detalhe de um campo invalido",
                "properties": {
                    "field": {
                        "type": "string",
                        "example": "itens[0].quantidade"
                    },
                    "issue": {
                        "type": "string",
                        "example": "deve ser maior que zero"
                    }
                },
                "required": [
                    "field",
                    "issue"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "description": "Formato padr\u00e3o de erro da API",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "ESTOQUE_INSUFICIENTE"
                    },
                    "message": {
                        "type": "string",
                        "example": "N\u00e3o ha quantidade suficiente para um ou mais itens."
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-05T12:00:00Z"
                    },
                    "path": {
                        "type": "string",
                        "example": "/pedidos"
                    },
                    "requestId": {
                        "type": "string",
                        "example": "3f1a2b4c-5d6e-7f80-91a2-b3c4d5e6f708"
                    }
                },
                "required": [
                    "details",
                    "error",
                    "message",
                    "timestamp"
                ]
            },
            "UnidadeResponse": {
                "type": "object",
                "description": "Unidade da rede",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "cidade": {
                        "type": "string"
                    },
                    "uf": {
                        "type": "string"
                    },
                    "tipoOperacao": {
                        "type": "string",
                        "enum": [
                            "COMPLETA",
                            "REDUZIDA"
                        ]
                    },
                    "ativa": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "ativa",
                    "cidade",
                    "id",
                    "nome",
                    "tipoOperacao",
                    "uf"
                ]
            },
...
            "ProdutoRequest": {
                "type": "object",
                "description": "Dados de um produto",
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Tapioca de queijo coalho",
                        "maxLength": 120,
                        "minLength": 2
                    },
                    "descricao": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    },
                    "categoria": {
                        "type": "string",
                        "example": "TAPIOCA",
                        "maxLength": 60,
                        "minLength": 0
                    },
                    "precoBase": {
                        "type": "number",
                        "description": "Pre\u00e7o sugerido pela matriz. Cada unidade pode praticar outro no cardapio.",
                        "example": 12.9,
                        "minimum": 0.0
                    },
                    "sazonal": {
                        "type": "boolean",
                        "description": "Produto vendido so em epoca especifica, tipo o periodo junino"
                    }
                },
                "required": [
                    "categoria",
                    "nome",
                    "precoBase",
                    "sazonal"
                ]
            },
            "ProdutoResponse": {
                "type": "object",
                "description": "Produto do catalogo da rede",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "categoria": {
                        "type": "string"
                    },
                    "precoBase": {
                        "type": "number"
                    },
                    "sazonal": {
                        "type": "boolean"
                    },
                    "ativo": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "ativo",
                    "categoria",
                    "id",
                    "nome",
                    "precoBase",
                    "sazonal"
                ]
            },
...
            "PaginaResponseProdutoResponse": {
                "type": "object",
                "description": "Listagem paginada",
                "properties": {
                    "conteudo": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProdutoResponse"
                        }
                    },
                    "pagina": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "limite": {
                        "type": "integer",
                        "format": "int32",
                        "example": 10
                    },
                    "totalItens": {
                        "type": "integer",
                        "format": "int64",
                        "example": 42
                    },
                    "totalPaginas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 5
                    },
                    "primeira": {
                        "type": "boolean"
                    },
                    "ultima": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "conteudo",
                    "limite",
                    "pagina",
                    "primeira",
                    "totalItens",
                    "totalPaginas",
                    "ultima"
                ]
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Cole aqui o accessToken devolvido por POST /auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}